Add feature configuration
Many features may better be handled by an application rather than by a
library, yet for less sophisticated clients it is better if a library
implements those.

A common solution is to implement optional features as loadable
modules (e.g., a collection of hooks and some state), which is also a
nice way to structure them. But module APIs tend to be restrictive, or
ad hoc (aiming to cover just the desired modules), or flexible and
reminiscent of "Greenspun's tenth rule", particularly when done in C.

The solution chosen here for now is to introduce configurable flags:
an application can enable or disable features that way, while the
control flow inside the library should be more explicit and not
restrictive.